refactor: step2 미션 범위에 맞게 수정
