--- tomcat/src/main/java/org/apache/coyote/controller/Controller.java
@@ -1,9 +1,0 @@
-package org.apache.coyote.controller;
-
-import org.apache.coyote.request.HttpRequest;
-import org.apache.coyote.response.HttpResponse;
-
-public interface Controller {
-
-    void service(HttpRequest request, HttpResponse response) throws Exception;
-}
