--- tomcat/src/main/java/org/apache/coyote/mapper/RequestValue.java
@@ -1,6 +1,0 @@
-package org.apache.coyote.mapper;
-
-import org.apache.coyote.http11.HttpMethod;
-
-public record RequestValue(HttpMethod method, String targetPath) {
-}
